+++ app/config.py
@@ -20,6 +20,7 @@ class AppConfig(BaseModel):
     save_to: str = "media"
     subdir: str = "scanner"
     resolution: int = 300
+    printer_name: str = ""
     telegram: TelegramConfig = Field(default_factory=TelegramConfig)